fix all cargo clippy
