+++ dcrd/src/chaincfg/chainhash/error.rs
