+++ dcrd/src/dcrjson/error.rs
