+++ dcrd/src/dcrutil/amount/error.rs
