+++ dcrd/src/dcrutil/amount/test.rs
