+++ dcrd/src/dcrutil/mod.rs
@@ -1,5 +1,5 @@
 pub mod amount;
 mod app_data;
-mod tests;
+mod test;
 
 pub use app_data::get_app_data_dir;
