+++ dcrd/src/dcrutil/test.rs
@@ -14,18 +14,12 @@ mod app_data_dir {
 
         assert_eq!(
             Some(home_dir),
-            crate::dcrutil::app_data::get_app_data_dir("myapp".into(), false)
+            crate::dcrutil::app_data::get_app_data_dir("myapp", false)
         );
 
-        assert_eq!(
-            None,
-            crate::dcrutil::app_data::get_app_data_dir("".into(), false)
-        );
+        assert_eq!(None, crate::dcrutil::app_data::get_app_data_dir("", false));
 
-        assert_eq!(
-            None,
-            crate::dcrutil::app_data::get_app_data_dir(".".into(), false)
-        );
+        assert_eq!(None, crate::dcrutil::app_data::get_app_data_dir(".", false));
     }
 
     #[test]
@@ -51,7 +45,7 @@ mod app_data_dir {
 
         assert_eq!(
             Some(home_dir),
-            crate::dcrutil::app_data::get_app_data_dir(&mut "myapp".into(), false)
+            crate::dcrutil::app_data::get_app_data_dir(&mut "myapp", false)
         )
     }
 
@@ -68,7 +62,7 @@ mod app_data_dir {
 
         assert_eq!(
             Some(home_dir),
-            crate::dcrutil::app_data::get_app_data_dir(&mut "myapp".into(), true)
+            crate::dcrutil::app_data::get_app_data_dir(&mut "myapp", true)
         )
     }
 
@@ -84,7 +78,7 @@ mod app_data_dir {
 
         assert_eq!(
             Some(home_dir),
-            crate::dcrutil::app_data::get_app_data_dir(&mut "myapp".into(), false)
+            crate::dcrutil::app_data::get_app_data_dir(&mut "myapp", false)
         )
     }
 
@@ -100,7 +94,7 @@ mod app_data_dir {
 
         assert_eq!(
             Some(home_dir),
-            crate::dcrutil::app_data::get_app_data_dir(&mut "myapp".into(), false)
+            crate::dcrutil::app_data::get_app_data_dir(&mut "myapp", false)
         )
     }
 }
