+++ dcrd/src/rpcclient/error.rs
