+++ dcrd/src/rpcclient/mod.rs
@@ -1,11 +1,4 @@
 #![cfg(feature = "rpcclient")]
-pub mod client;
-pub mod connection;
 pub(crate) mod constants;
-mod errors;
-mod infrastructure;
-pub use self::errors::RpcClientError;
-mod chain_command;
-mod chain_notification;
-pub mod notify;
-pub mod tests;
+pub mod error;
+pub mod test;
