+++ dcrd/src/rpcclient/test.rs
